// src/core/interface/api-endpoint.interface.ts
export enum ApiEndPoint {
  MARINE_VESSEL = 'marine-vessel-applications',
  NCEC = 'ncec-applications',
  NCRC = 'ncrc-applications',
  EP_APPLICATION = 'exchange-program-applications',
  EP_REQUEST = 'exchange-program-requests',
  REGISTRATION_CERTIFICATE = 'registration-certificate-applications',
  ADVERT_APPLICATION = 'advert-applications',
  ADVERT = 'adverts',
  TWP_APPLICATION = 'temporary-work-permit-applications',
  TWP_REQUEST = 'temporary-work-permit-requests',
  FAQ = 'faqs',
  DOCUMENT = 'generic-documents',
  STAFF = 'staffs',
  BASE_RECORD = 'base-records',
  USERS = 'users',
  SERVICE_EXECUTED = 'service-executed',
  SERVICE = 'services',
  SHAREHOLDER = 'ownership-shareholder',
  EQUIPMENT = 'equipments',
  RAW_MATERIAL = 'raw-materials',
  QUALITY_ASSURANCE = 'quality-assurance',
  SUBSIDIARY = 'Subsidiaries',
  NUPRC = 'nuprc',
  NC_POLICY = 'nc-policies',
  ORGANOGRAM = 'organogram',
  OEMS = 'oems',
  FACILITY = 'facilities',
  AUTH = 'auth',
  ROLE = 'roles',
  PERMISSION = 'permissions',
  NCTRC = 'nctrc-applications',
  NCTRC_COURSE = 'nctrc-courses',
  NCTRC_TYPE = 'nctrc-types',
  NCTRC_MATRIX_CRITERIA = 'nctrc-matrix-criteria',
  ACCOUNTS = 'accounts',
  ACCOUNTS_PUBLIC = 'accounts-public',
  ACCOUNTS_AUDITOR = 'account-auditors',
  EQ_APPLICATION = 'expatriate-quota-applications',
  EQ_REQUEST = 'expatriate-quota-requests',
  WORKFLOW = 'workflows',
  WHISTLE_BLOWER = 'whistle-blower',
  OWNERSHIP_DOCUMENT = 'ownership-documents',
  BULK_UPLOAD = 'bulk-uploads',
  EMAIL_TEMPLATE = 'mail',
  // PROMOTION START
  PROMOTION_PRODUCT = 'promotion/product',
  PROMOTION_PRODUCT_APPLICATION = 'promotion/product-applications',
  PROMOTION_PRODUCT_ADVERTS = 'promotion/product-advert',
  PROMOTION_ADEVRT_PACKAGES = 'promotion/advert-packages',
  PROMOTION_SERVICE = 'promotion/service',
  PROMOTION_SERVICE_ADVERTS = 'promotion/service-advert',
  PROMOTION_SERVICE_APPLICATION = 'promotion/service-applications',
  PROMOTION_SERVICE_PUBLIC = 'promotion/service-public',
  // PROMoTION END
  TENDER_APPLICATION = 'tender-applications',
  TENDER_APPLICATION_STAGE = 'tender-application-stages',
  TENDER_OPPORTUNITY = 'tender-opportunities',
  TENDER_CONTRACT_ADVERT = 'tender-contract-adverts',
  TENDER_CONTRACT_STRATEGY = 'tender-contract-strategy',

  SERVICE_CODE = 'service-codes',
  SERVICE_CODE_APPLICATION = 'service-code-apps',
  GUIDELINE = 'guidelines',
  NCDF = 'ncdf',
  JOB_FORECAST_APPLICATION = 'job-forecast-applications',
  JOB_FORECAST_REQUEST = 'job-forecast-requests',
  JOB_FORECAST_MATERIAL_REQUEST = 'job-forecast-material-requests',

  // R&D
  RESEARCH_APPLICATION = 'research-applications',
  RESEARCH_APPLICATION_EVALUATION = 'research-applications-evaluations',

  // SERVICE UTILIZATION
  SERVICE_UTILIZATION_APPLICATION = 'service-utilization-applications',
  SERVICE_UTILIZATION_REQUEST = 'service-utilization-requests',
  SERVICE_UTILIZATION_REPORT = 'service-utilization-reports',
  MVC_REPORT_ITEM = 'marine-vessel-report-items',
  INDIVIDUAL_HCD_TRAINING_PROGRAM = 'individual-hcd-training-program',
  INDIVIDUAL_HCD_TRAINING_SKILL_BASE = 'individual-hcd-training-skill-base',
}