// src/iam/enum/permission.enum.ts
export enum PermisionActionTypeEnum {
  READ = 'read',
  UPDATE = 'update',
  DELETE = 'delete',
  CREATE = 'create',
  MANAGE = 'manage',
  READ_REPORT = 'read-report',
  ARCHIVE = 'archived',
  ACTIVATION = 'activation',
  PASSWORD = 'password',
  WORKFLOW_GROUP = 'workflow-group',
  REPORT = 'report',
  DOWNLOAD = 'download',
  CREATE_DOCUMENTS = 'create-documents',
  RESEND_ACKNOWLEDGEMENT = 'resend-acknowledgement',
  EVALUATION = 'evaluate',
  REASSIGN = 'reassign',
}

export enum PermisionSubjectTypeEnum {
  USER = 'user',
  ALL = 'all',
  AUDIT_LOGS = 'audit-logs',
  INDIVIDUAL = 'individual',
  NCRC = 'ncrc',
  NCEC = 'ncec',
  NCTRC = 'nctrc',
  NCTRC_TRAINING_COURSE = 'nctrc-training-course',
  NCTRC_MATRIX_CRITERIA = 'nctrc-matrix-criteria',
  NCTRC_SPECIALIZATION = 'nctrc-area-of-specialization',
  BASE_RECORD = 'base-record',
  EXCHANGE_PROGRAM = 'exchange-program',
  MARINE_VESSEL = 'marine-vessel',
  ADVERT = 'advert',
  DOCUMENT = 'document',
  EQ = 'eq',
  TEMPORARY_WORK_PERMIT = 'temporary-work-permit',
  MARINE_VESSEL_REPORT_QUARTERLY = 'marine-vessel-quarterly-report',
  WORKFLOW = 'workflow',
  SERVICE_CODE = 'service-code',
  SETTINGS = 'settings',
  PAYMENT = 'payment',
}